+++ src/commands/spconfig-export/SPConfigExporter.ts
@@ -1,5 +1,4 @@
 import * as vscode from 'vscode';
-
 import * as fs from 'fs';
 import path = require('path');